+++ src/System.Windows.Forms.DataVisualization/General/DataManipulator.cs
@@ -2982,7 +2982,7 @@ namespace System.Windows.Forms.DataVisualization.Charting
 			int	numberOfYValues = 0;
 			foreach(Series series in inputSeries)
 			{
-				numberOfYValues = (int)Math.Max(numberOfYValues, series.YValuesPerPoint);
+				numberOfYValues = Math.Max(numberOfYValues, series.YValuesPerPoint);
 			}
 
 			// Allocate memory for the result array for X and each Y values
